wscript: make it build in windows native enviroment
When collecting build specs database a forward slash was added to the
uid to make it into a path name, which is used as key of the build spec
dictionary.

Later when tring to retrive build spec from the dictionary during
'waf configure`, the target uid is generated with `os.path.normpath` and
in windows native environment two backward slashes are added to the uid,
and then no valid spec could be found.

This is fixed by removing the forward slash when building the build spec
database.